Implement the Http Entity

+++ src/main/java/com/tanuj/journalApp/controller/JournalEntryControllerV2.java
@@ -4,13 +4,13 @@ import com.tanuj.journalApp.entity.JournalEntry;
 import com.tanuj.journalApp.service.JournalEntryService;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @RestController
 @RequestMapping("/journal")
@@ -20,39 +20,59 @@ public class JournalEntryControllerV2 {
     private JournalEntryService JournalEntryService;
 
     @GetMapping("get-all")
-    public List<JournalEntry> getAll() {
-        return JournalEntryService.getAll();
+    public ResponseEntity<List<JournalEntry>> getAll() {
+        List<JournalEntry> JournalEntry = JournalEntryService.getAll();
+        if(JournalEntry != null && !JournalEntry.isEmpty()){
+            return new ResponseEntity<>(JournalEntry, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @PostMapping("add-journal")
-    public JournalEntry addJournal(@RequestBody JournalEntry myEntry) {
-        myEntry.setDate(LocalDateTime.now());
-        JournalEntryService.saveEntry(myEntry);
-        return myEntry;
+    public ResponseEntity<JournalEntry> addJournal(@RequestBody JournalEntry myEntry) {
+
+        try {
+            myEntry.setDate(LocalDateTime.now());
+            JournalEntryService.saveEntry(myEntry);
+            return new ResponseEntity<>(myEntry, HttpStatus.CREATED);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+
+        }
     }
 
     @GetMapping("id/{myId}")
-    public JournalEntry getJournalByID(@PathVariable ObjectId myId) {
-        return JournalEntryService.findById(myId).orElse(null);
+    public ResponseEntity<JournalEntry> getJournalByID(@PathVariable ObjectId myId) {
+        Optional<JournalEntry> JournalEntry = JournalEntryService.findById(myId);
 
+        if (JournalEntry.isPresent()) {
+            return new ResponseEntity<>(JournalEntry.get(), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @DeleteMapping("id/{myId}")
-    public boolean deleteEntry(@PathVariable ObjectId myId) {
-        JournalEntryService.deleteEntry(myId);
-        return true;
+    public ResponseEntity<JournalEntry> deleteEntry(@PathVariable ObjectId myId) {
+        Optional<JournalEntry> JournalEntry = JournalEntryService.findById(myId);
+
+        if(JournalEntry.isPresent()){
+            JournalEntryService.deleteEntry(myId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     @PutMapping("id/{id}")
-    public JournalEntry updateEntrybyId(@PathVariable ObjectId id, @RequestBody JournalEntry myEntry) {
-        JournalEntry find= JournalEntryService.findById(id).orElse(null);
+    public ResponseEntity<JournalEntry> updateEntrybyId(@PathVariable ObjectId id, @RequestBody JournalEntry myEntry) {
+        JournalEntry find = JournalEntryService.findById(id).orElse(null);
 
-        if(find != null){
+        if (find != null) {
             find.setTitle(myEntry.getTitle() != null && !myEntry.getTitle().isEmpty() ? myEntry.getTitle() : find.getTitle());
             find.setContent(myEntry.getContent() != null && !myEntry.getContent().isEmpty() ? myEntry.getContent() : find.getContent());
+            JournalEntryService.saveEntry(find);
+            return new ResponseEntity<>(find,HttpStatus.OK);
         }
 
-        JournalEntryService.saveEntry(find);
-        return find;
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
